+++ .gitignore
@@ -124,7 +124,6 @@ dmypy.json
 
 # Editors
 .vscode/
-*.swp
 
 # Dask
 dask-worker-space/
@@ -135,10 +134,6 @@ dask-worker-space/
 # Generic data directories
 **/data/*
 
-# Nohup
-nohup.out
-
 # Modeling
 checkpoint/
 checkpoints/
-
