+++ spouse/.gitignore
@@ -1,3 +1,2 @@
 # Stored data
 *.pkl
-trained_spouse_model
